requirements.txt
redis
pyignite
python-dotenv